Expose audio play API as a service
The audio API can be used asynchronously: we want to be able to start
the playback without stopping the calling thread (e.g. from Aseba).

To do so a service is exposed via two topics: one is used to send a
request to the audio file, and the other contains the reply.